Cleaned up package
